Implement EntityAutoWalk()
WOW!!!!! There's a lot to cover here.

Part of games like Shining Force and Fire Emblem are enemies who move
automatically from spot to spot, pathfinding around various things.

There's probably a ton of ways to do this, but here's how I did it:

1. Keep track of your origin point
2. Use Pathfinding to find a path to your destination
3. Build up that path -- from that path, we will construct a series of
steps with all the information that we need.

For each step:
- Direction to be walking
- Coordinate to move to

  0 1 2 3
  1 S
  2
  3     T

A path might look something like this:

  0 1 2 3
  1 S
  2 |
  3 L _ T

So going the steps would be:
1. Down - (1,2)
2. Down - (1,3)
3. Right - (2,3)
4. Right - (3,3)

4. Construct a series of Timers in a loop
- Each timer will begin at an offest time equal to
 (Step Number - 1) * ENTITY_WALK_SPEED
- This ensures that we hae a cascading series of Timer Moves

1. does stuf...
  2. does stuff after a little bit...
     3. does stuff after even longer...
	4. finally does stuff after even longer...

- This allows all the animations to be smoothed
5. At the very end, we will reset the gridX, gridY of the entity
and have them face south once more